+++ src/devices/frient.ts
@@ -1,80 +1,9 @@
-import {Zcl} from 'zigbee-herdsman';
-
 import {develcoModernExtend} from '../lib/develco';
-import * as exposes from '../lib/exposes';
 import {battery, electricityMeter, onOff, ota} from '../lib/modernExtend';
 import * as reporting from '../lib/reporting';
-import {DefinitionWithExtend, Fz, ModernExtend, Tz} from '../lib/types';
-import * as utils from '../lib/utils';
-
-const e = exposes.presets;
-const ea = exposes.access;
+import {DefinitionWithExtend} from '../lib/types';
 
 // NOTE! Develco and Frient is the same company, therefore we use develco specific things in here.
-
-// frient/develco specific constants
-const manufacturerOptions = {manufacturerCode: Zcl.ManufacturerCode.DEVELCO};
-
-function current_summation() {
-    return {
-        isModernExtend: true,
-        toZigbee: [
-            {
-                key: ['current_summation'],
-                convertSet: async (entity, key, value, meta) => {
-                    await entity.write('seMetering', {develcoCurrentSummation: value}, utils.getOptions(meta.mapped, entity));
-                    return {state: {current_summation: value}};
-                },
-            } satisfies Tz.Converter,
-        ],
-        exposes: [
-            e
-                .numeric('current_summation', ea.SET)
-                .withDescription('Current summation value sent to the display. e.g. 570 = 0,570 kWh')
-                .withValueMin(0)
-                .withValueMax(268435455),
-        ],
-    } satisfies ModernExtend;
-}
-
-function pulse_configuration() {
-    return {
-        isModernExtend: true,
-        fromZigbee: [
-            {
-                cluster: 'seMetering',
-                type: ['attributeReport', 'readResponse'],
-                convert: (model, msg, publish, options, meta) => {
-                    const result: Record<string, unknown> = {};
-                    if (msg.data?.develcoPulseConfiguration) {
-                        result[utils.postfixWithEndpointName('pulse_configuration', msg, model, meta)] = msg.data['develcoPulseConfiguration'];
-                    }
-
-                    return result;
-                },
-            } satisfies Fz.Converter,
-        ],
-        toZigbee: [
-            {
-                key: ['pulse_configuration'],
-                convertSet: async (entity, key, value, meta) => {
-                    await entity.write('seMetering', {develcoPulseConfiguration: value}, utils.getOptions(meta.mapped, entity));
-                    return {readAfterWriteTime: 200, state: {pulse_configuration: value}};
-                },
-                convertGet: async (entity, key, meta) => {
-                    await entity.read('seMetering', ['develcoPulseConfiguration'], manufacturerOptions);
-                },
-            } satisfies Tz.Converter,
-        ],
-        exposes: [
-            e
-                .numeric('pulse_configuration', ea.ALL)
-                .withValueMin(0)
-                .withValueMax(65535)
-                .withDescription('Pulses per kwh. Default 1000 imp/kWh. Range 0 to 65535'),
-        ],
-    } satisfies ModernExtend;
-}
 
 const definitions: DefinitionWithExtend[] = [
     {
@@ -88,8 +17,8 @@ const definitions: DefinitionWithExtend[] = [
             battery(),
             develcoModernExtend.addCustomClusterManuSpecificDevelcoGenBasic(),
             develcoModernExtend.readGenBasicPrimaryVersions(),
-            pulse_configuration(),
-            current_summation(),
+            develcoModernExtend.pulse_configuration(),
+            develcoModernExtend.current_summation(),
         ],
 
         configure: async (device, coordinatorEndpoint, logger) => {
